fix the mistakes

+++ src/controllers/authController.js
@@ -30,7 +30,7 @@ export const loginUser = async (req, res) => {
   if (!isValidPassword) {
     throw createHttpError(401, 'Invalid credentials');
   }
-  await Session.deleteOne(user._id);
+  await Session.deleteOne({ userId: user._id });
   const newSession = await createSession(user._id);
   setSessionCookies(res, newSession);
   res.status(200).json(user);
